# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/powerbanks": {
            "get": {
                "description": "Возвращает все повербанки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "powerbanks"
                ],
                "summary": "Возвращает список повербанков",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PowerbankListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создаёт повербанк с заданными current_station_id и status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "powerbanks"
                ],
                "summary": "Создаёт новый повербанк",
                "parameters": [
                    {
                        "description": "Powerbank Data",
                        "name": "powerbank",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreatePowerbankRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.PowerbankResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/powerbanks/{id}": {
            "get": {
                "description": "Возвращает повербанк по переданному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "powerbanks"
                ],
                "summary": "Получает повербанк по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Powerbank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PowerbankResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные повербанка. Обновляются только переданные поля.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "powerbanks"
                ],
                "summary": "Обновляет повербанк по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Powerbank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Powerbank Data",
                        "name": "powerbank",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdatePowerbankRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PowerbankResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет повербанк с заданным идентификатором",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "powerbanks"
                ],
                "summary": "Удаляет повербанк по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Powerbank ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Powerbank": {
            "type": "object",
            "properties": {
                "current_station_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/data.PowerbankStatus"
                }
            }
        },
        "data.PowerbankStatus": {
            "type": "string",
            "enum": [
                "rented",
                "available",
                "charging"
            ],
            "x-enum-varnames": [
                "PowerbankStatusRented",
                "PowerbankStatusAvailable",
                "PowerbankStatusCharging"
            ]
        },
...
        "main.CreatePowerbankRequest": {
            "type": "object",
            "properties": {
                "current_station_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "main.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.PowerbankListResponse": {
            "type": "object",
            "properties": {
                "powerbanks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.Powerbank"
                    }
                }
            }
        },
        "main.PowerbankResponse": {
            "type": "object",
            "properties": {
                "powerbank": {
                    "$ref": "#/definitions/data.Powerbank"
                }
            }
        },
...
        "main.UpdatePowerbankRequest": {
            "type": "object",
            "properties": {
                "current_station_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}